Create wds-banner.js
this will be shown on the wds page. And because there is some legacy z-index on the nav that for some unknown/defies CSS reason won't listen to monetates overrides in css so I have to wait till the js is done firing (hence the wait function on line 74) and then have to override it calling the actual tag rather than the class. So help me god I cannot wait till everything is on gatsby.