Create rich-pdp-hardcover.js
This was developed by a marketing contractor and brandon but wanted to push it through eng because it's the right thing to do. This adds carousel functionality to the additional content under the PDP buy box. The top minified js is Jquery, and the next one is owl-carousel minified. The codepen for look,feel, and functionality is here: https://codepen.io/brandon-artifact/pen/GRqQMYx